Node & MiniMaxIA
Put some basic functions like AddNode in the nextNodes list and setscore which is a attribute evaluating the position of the plateau.
nextNodes is the sons of the current node.

MiniMaxIA has no nex attributs but has a CalculScore function for one node. It will be used to the minimax algorithm